membuat kondisional if di method PUT

+++ src/main/java/ws/c/project04/wscproject04/ProductServiceController.java
@@ -46,6 +46,9 @@ public class ProductServiceController {
    
    @RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
    public ResponseEntity<Object> updateProduct(@PathVariable("id")String id, @RequestBody Product product){
+       if(!productRepo.containsKey(id)){
+           return new ResponseEntity<>("Product key is wrong", HttpStatus.OK);
+       }
        productRepo.remove(id);
        product.setId(id);
        productRepo.put(id, product);
